Removed a feature providing ';' after each line
The reason is, neither every line was intended to end with a semicolon nor is able to, without breaking the code logic.Altough a redundant semicolon stands for an empty statement, wich does not usually affect anything, in some places such as after loops of if's such a statement is replacing the body thus breaking logic.
This could be fixed with regex-checks for the each line, howewer, the loss of performance and the code robustness makes it too expencive. This feature considered rather useless.

Also, a newline character is now appended at the end of each line, this leads to more meaningfull output from compiler, because the index of each line in the source file corresponds to order number of an actual user input lines.