Added method setJdkPath()
jdkPath variable is now initialized there and not in the constructor.
After an attempt to locate jdk using JAVA_HOME, if it is not set,
application lets user to input a path to jdk manually

minor fix: in method run() every new line of input is checked to not be empty.
Empty lines are simply ignored, this is made because isCommand() method used to throw an exception when called on an empty string.